refactor: split synapse analysis modules for maintainability (#802)
Split the two largest files in src/analysis/synapse/:

gpu_evaluation.rs (988 → 244 lines):
- relu_evaluation.rs: ReLU candidate evaluation
- activation_subset_evaluation.rs: split-error subset evaluation
- activation_evaluation.rs: per-spec activation candidate evaluation
- gpu_evaluation.rs: batched orchestration + re-exports

mod.rs (945 → 155 lines):
- orchestration.rs: core analysis pipeline coordination
- metadata.rs: lock-free atomic metadata and result merging
- results.rs: result finalisation and assembly
- tests.rs: unit tests
- mod.rs: thin re-export layer + public API

All existing tests pass without modification. Public API unchanged.
